Cambio login

+++ src/main/java/py/com/sgipy/miesys/controllers/CiudadJpaController.java
@@ -10,11 +10,13 @@ import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
-import py.com.sgipy.miesys.entities.Direccion;
+import py.com.sgipy.miesys.entities.Departamento;
+import py.com.sgipy.miesys.entities.Barrio;
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
+import py.com.sgipy.miesys.controllers.exceptions.IllegalOrphanException;
 import py.com.sgipy.miesys.controllers.exceptions.NonexistentEntityException;
 import py.com.sgipy.miesys.entities.Ciudad;
 
@@ -34,27 +36,36 @@ public class CiudadJpaController implements Serializable {
     }
 
     public void create(Ciudad ciudad) {
-        if (ciudad.getDireccionList() == null) {
-            ciudad.setDireccionList(new ArrayList<Direccion>());
+        if (ciudad.getBarrioList() == null) {
+            ciudad.setBarrioList(new ArrayList<Barrio>());
         }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            List<Direccion> attachedDireccionList = new ArrayList<Direccion>();
-            for (Direccion direccionListDireccionToAttach : ciudad.getDireccionList()) {
-                direccionListDireccionToAttach = em.getReference(direccionListDireccionToAttach.getClass(), direccionListDireccionToAttach.getDireccion());
-                attachedDireccionList.add(direccionListDireccionToAttach);
+            Departamento departamento = ciudad.getDepartamento();
+            if (departamento != null) {
+                departamento = em.getReference(departamento.getClass(), departamento.getDepartamento());
+                ciudad.setDepartamento(departamento);
             }
-            ciudad.setDireccionList(attachedDireccionList);
+            List<Barrio> attachedBarrioList = new ArrayList<Barrio>();
+            for (Barrio barrioListBarrioToAttach : ciudad.getBarrioList()) {
+                barrioListBarrioToAttach = em.getReference(barrioListBarrioToAttach.getClass(), barrioListBarrioToAttach.getBarrio());
+                attachedBarrioList.add(barrioListBarrioToAttach);
+            }
+            ciudad.setBarrioList(attachedBarrioList);
             em.persist(ciudad);
-            for (Direccion direccionListDireccion : ciudad.getDireccionList()) {
-                Ciudad oldCiudadOfDireccionListDireccion = direccionListDireccion.getCiudad();
-                direccionListDireccion.setCiudad(ciudad);
-                direccionListDireccion = em.merge(direccionListDireccion);
-                if (oldCiudadOfDireccionListDireccion != null) {
-                    oldCiudadOfDireccionListDireccion.getDireccionList().remove(direccionListDireccion);
-                    oldCiudadOfDireccionListDireccion = em.merge(oldCiudadOfDireccionListDireccion);
+            if (departamento != null) {
+                departamento.getCiudadList().add(ciudad);
+                departamento = em.merge(departamento);
+            }
+            for (Barrio barrioListBarrio : ciudad.getBarrioList()) {
+                Ciudad oldCiudadOfBarrioListBarrio = barrioListBarrio.getCiudad();
+                barrioListBarrio.setCiudad(ciudad);
+                barrioListBarrio = em.merge(barrioListBarrio);
+                if (oldCiudadOfBarrioListBarrio != null) {
+                    oldCiudadOfBarrioListBarrio.getBarrioList().remove(barrioListBarrio);
+                    oldCiudadOfBarrioListBarrio = em.merge(oldCiudadOfBarrioListBarrio);
                 }
             }
             em.getTransaction().commit();
@@ -65,36 +76,56 @@ public class CiudadJpaController implements Serializable {
         }
     }
 
-    public void edit(Ciudad ciudad) throws NonexistentEntityException, Exception {
+    public void edit(Ciudad ciudad) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
             Ciudad persistentCiudad = em.find(Ciudad.class, ciudad.getCiudad());
-            List<Direccion> direccionListOld = persistentCiudad.getDireccionList();
-            List<Direccion> direccionListNew = ciudad.getDireccionList();
-            List<Direccion> attachedDireccionListNew = new ArrayList<Direccion>();
-            for (Direccion direccionListNewDireccionToAttach : direccionListNew) {
-                direccionListNewDireccionToAttach = em.getReference(direccionListNewDireccionToAttach.getClass(), direccionListNewDireccionToAttach.getDireccion());
-                attachedDireccionListNew.add(direccionListNewDireccionToAttach);
-            }
-            direccionListNew = attachedDireccionListNew;
-            ciudad.setDireccionList(direccionListNew);
-            ciudad = em.merge(ciudad);
-            for (Direccion direccionListOldDireccion : direccionListOld) {
-                if (!direccionListNew.contains(direccionListOldDireccion)) {
-                    direccionListOldDireccion.setCiudad(null);
-                    direccionListOldDireccion = em.merge(direccionListOldDireccion);
+            Departamento departamentoOld = persistentCiudad.getDepartamento();
+            Departamento departamentoNew = ciudad.getDepartamento();
+            List<Barrio> barrioListOld = persistentCiudad.getBarrioList();
+            List<Barrio> barrioListNew = ciudad.getBarrioList();
+            List<String> illegalOrphanMessages = null;
+            for (Barrio barrioListOldBarrio : barrioListOld) {
+                if (!barrioListNew.contains(barrioListOldBarrio)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Barrio " + barrioListOldBarrio + " since its ciudad field is not nullable.");
                 }
             }
-            for (Direccion direccionListNewDireccion : direccionListNew) {
-                if (!direccionListOld.contains(direccionListNewDireccion)) {
-                    Ciudad oldCiudadOfDireccionListNewDireccion = direccionListNewDireccion.getCiudad();
-                    direccionListNewDireccion.setCiudad(ciudad);
-                    direccionListNewDireccion = em.merge(direccionListNewDireccion);
-                    if (oldCiudadOfDireccionListNewDireccion != null && !oldCiudadOfDireccionListNewDireccion.equals(ciudad)) {
-                        oldCiudadOfDireccionListNewDireccion.getDireccionList().remove(direccionListNewDireccion);
-                        oldCiudadOfDireccionListNewDireccion = em.merge(oldCiudadOfDireccionListNewDireccion);
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            if (departamentoNew != null) {
+                departamentoNew = em.getReference(departamentoNew.getClass(), departamentoNew.getDepartamento());
+                ciudad.setDepartamento(departamentoNew);
+            }
+            List<Barrio> attachedBarrioListNew = new ArrayList<Barrio>();
+            for (Barrio barrioListNewBarrioToAttach : barrioListNew) {
+                barrioListNewBarrioToAttach = em.getReference(barrioListNewBarrioToAttach.getClass(), barrioListNewBarrioToAttach.getBarrio());
+                attachedBarrioListNew.add(barrioListNewBarrioToAttach);
+            }
+            barrioListNew = attachedBarrioListNew;
+            ciudad.setBarrioList(barrioListNew);
+            ciudad = em.merge(ciudad);
+            if (departamentoOld != null && !departamentoOld.equals(departamentoNew)) {
+                departamentoOld.getCiudadList().remove(ciudad);
+                departamentoOld = em.merge(departamentoOld);
+            }
+            if (departamentoNew != null && !departamentoNew.equals(departamentoOld)) {
+                departamentoNew.getCiudadList().add(ciudad);
+                departamentoNew = em.merge(departamentoNew);
+            }
+            for (Barrio barrioListNewBarrio : barrioListNew) {
+                if (!barrioListOld.contains(barrioListNewBarrio)) {
+                    Ciudad oldCiudadOfBarrioListNewBarrio = barrioListNewBarrio.getCiudad();
+                    barrioListNewBarrio.setCiudad(ciudad);
+                    barrioListNewBarrio = em.merge(barrioListNewBarrio);
+                    if (oldCiudadOfBarrioListNewBarrio != null && !oldCiudadOfBarrioListNewBarrio.equals(ciudad)) {
+                        oldCiudadOfBarrioListNewBarrio.getBarrioList().remove(barrioListNewBarrio);
+                        oldCiudadOfBarrioListNewBarrio = em.merge(oldCiudadOfBarrioListNewBarrio);
                     }
                 }
             }
@@ -115,7 +146,7 @@ public class CiudadJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -127,10 +158,21 @@ public class CiudadJpaController implements Serializable {
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The ciudad with id " + id + " no longer exists.", enfe);
             }
-            List<Direccion> direccionList = ciudad.getDireccionList();
-            for (Direccion direccionListDireccion : direccionList) {
-                direccionListDireccion.setCiudad(null);
-                direccionListDireccion = em.merge(direccionListDireccion);
+            List<String> illegalOrphanMessages = null;
+            List<Barrio> barrioListOrphanCheck = ciudad.getBarrioList();
+            for (Barrio barrioListOrphanCheckBarrio : barrioListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Ciudad (" + ciudad + ") cannot be destroyed since the Barrio " + barrioListOrphanCheckBarrio + " in its barrioList field has a non-nullable ciudad field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            Departamento departamento = ciudad.getDepartamento();
+            if (departamento != null) {
+                departamento.getCiudadList().remove(ciudad);
+                departamento = em.merge(departamento);
             }
             em.remove(ciudad);
             em.getTransaction().commit();
